+++ spark/v3.3/spark/src/main/java/org/apache/iceberg/spark/source/EqualityDeleteRowReader.java
@@ -39,7 +39,7 @@ public class EqualityDeleteRowReader extends RowDataReader {
 
   @Override
   CloseableIterator<InternalRow> open(FileScanTask task) {
-    SparkDeleteFilter matches = new SparkDeleteFilter(task, table().schema(), expectedSchema, this);
+    SparkDeleteFilter matches = new SparkDeleteFilter(task.file().path().toString(), task.deletes(), expectedSchema);
 
     // schema or rows returned by readers
     Schema requiredSchema = matches.requiredSchema();
